+++ src/content_delegate.js
@@ -180,6 +180,27 @@ ContentDelegate.prototype.findAndStorePacerDocIds = function () {
   });
 };
 
+// If this is a docket query page, add RECAP email advertisement.
+ContentDelegate.prototype.addRecapEmailAdvertisement = async function () {
+  if (!(PACER.isBlankQueryReportUrl(this.url) || PACER.isManageAccountPage(this.url))) { return; };
+  let form; 
+
+  if (!PACER.hasFilingCookie(document.cookie)){ return; };
+   
+  if (PACER.isBlankQueryReportUrl(this.url)){
+    form = document.querySelector('form');
+    if (!document.querySelector('.recap-email-banner')){
+      form.appendChild(recapEmailBanner());
+    }
+  }else{
+    form = document.querySelector('#popupForm');
+    if(!document.querySelector('.recap-email-banner-full')){
+      form.after(recapEmailBanner('recap-email-banner-full'));
+    }
+  }
+  
+};
+
 // If this is a docket query page, ask RECAP whether it has the docket page.
 ContentDelegate.prototype.handleDocketQueryUrl = function () {
   if (!PACER.isDocketQueryUrl(this.url)) { return; };